% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facade.R
\name{add_facade_alt}
\alias{add_facade_alt}
\title{Add a facade to a tsg table (alternative way)}
\usage{
add_facade_alt(data, ...)
}
\arguments{
\item{data}{a \code{tsg} data frame}

\item{...}{List of supported facade items (see \code{\link[=add_facade]{add_facade()}})}
}
\value{
A \code{tsg} object with the specified facade settings applied as attributes.
}
\description{
This function adds a facade to a \code{tsg}, an alternative way to allow dynamic values and programmatic evaluation.
}
\examples{
person_record |>
 generate_frequency(sex) |>
 add_facade_alt(table.offsetRow = 2, table.offsetCol = 1)
}
