% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{add_missing_label}
\alias{add_missing_label}
\title{Add a Label for Missing Values in a Variable}
\usage{
add_missing_label(value, label_na = "Not reported", recode_na = "auto")
}
\arguments{
\item{value}{A vector (labelled, factor, or atomic) in which missing values should be labeled.}

\item{label_na}{Character. The label to assign to missing values (default is `"Not reported"`).}

\item{recode_na}{Character or `"auto"`. If the vector is `haven::labelled`, defines the numeric code used to represent missing values.
`"auto"` will automatically generate a suitable value beyond the current max code.}
}
\value{
A vector with missing values replaced and labeled appropriately. Preserves variable attributes.
}
\description{
Assigns a custom label for missing (`NA`) values in a variable. Handles labelled vectors (via `haven::labelled`),
factors (via `forcats::fct_na_value_to_level()`), and other vectors by replacing `NA` with a user-defined value or label.
}
