% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_modifiers.R
\name{add_source_note}
\alias{add_source_note}
\title{Add a source note attribute to a table}
\usage{
add_source_note(data, source_note)
}
\arguments{
\item{data}{A data frame, tibble, or \code{tsg} object to which a source note attribute will be added.}

\item{source_note}{The source note text to be added.}
}
\value{
The input data frame with an added source note attribute.
}
\description{
Add a source note attribute to a table
}
\examples{
add_source_note(
  dplyr::starwars,
  source_note = "Source: Star Wars API (SWAPI)."
)
}
