% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_modifiers.R
\name{add_table_subtitle}
\alias{add_table_subtitle}
\title{Add a subtitle attribute to a table}
\usage{
add_table_subtitle(data, subtitle)
}
\arguments{
\item{data}{A data frame, tibble, or \code{tsg} object to which a subtitle attribute will be added.}

\item{subtitle}{The subtitle text to be added.}
}
\value{
The input data frame with an added subtitle attribute.
}
\description{
Add a subtitle attribute to a table
}
\examples{
add_table_subtitle(
  dplyr::starwars,
  subtitle = "Star Wars Character Data"
)

}
