% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_modifiers.R
\name{add_table_title}
\alias{add_table_title}
\title{Add a title attribute to a table}
\usage{
add_table_title(data, title)
}
\arguments{
\item{data}{A data frame, tibble, or \code{tsg} object to which a title attribute will be added.}

\item{title}{The title text to be added.}
}
\value{
The input data frame with an added title attribute.
}
\description{
Add a title attribute to a table
}
\examples{
add_table_title(
  dplyr::starwars,
  title = "Star Wars Character Data"
)
}
