% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_modifiers.R
\name{convert_factor}
\alias{convert_factor}
\title{Convert labelled factors to regular factors}
\usage{
convert_factor(data)
}
\arguments{
\item{data}{}
}
\description{
Convert labelled factors to regular factors
}
