% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_template.R
\name{generate_template}
\alias{generate_template}
\title{Generate a template}
\usage{
generate_template(
  path,
  template = c("facade", "table-list"),
  which = c("xlsx", "pdf", "html")
)
}
\arguments{
\item{path}{A character string specifying the path where the template should be saved. If a directory is provided, the template will be saved with a default name based on the template type.}

\item{template}{A character string specifying the type of template to generate. Options are "facade" for a YAML facade template or "table-list" for an Excel table list template.}

\item{which}{A character string specifying the format of the template to generate. Options are "xlsx", "pdf", or "html". Defaults to "xlsx".}
}
\description{
Generate a template
}
