% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facade.R
\name{get_tsg_facade}
\alias{get_tsg_facade}
\title{Get a facade from the package or a file}
\usage{
get_tsg_facade(facade = "default", which = c("xlsx", "html", "docx", "pdf"))
}
\arguments{
\item{facade}{A character string specifying the name of the facade to retrieve. Defaults to "default". The facade is a YAML file that defines the styling and layout of the table}

\item{which}{A character string specifying the format of the facade to retrieve. Options are "xlsx", "pdf", or "html". Defaults to "xlsx".}
}
\value{
A list containing the facade settings for the specified format. The facade includes styling attributes such as font size, color, border styles, and background fills for different parts of the table.
}
\description{
Get a facade from the package or a file
}
\examples{

# Default facade
get_tsg_facade()

# Other built-in facade
get_tsg_facade("yolo")
}
