% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{gt_apply_spanners}
\alias{gt_apply_spanners}
\title{Apply spanner labels to a gt table based on separator-split column names}
\usage{
gt_apply_spanners(gt_tbl, col_names, sep, col_labels = NULL)
}
\description{
Apply spanner labels to a gt table based on separator-split column names
}
\keyword{internal}
