% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{gt_apply_text_decoration}
\alias{gt_apply_text_decoration}
\title{Apply text decoration (bold/italic/underline) to gt locations}
\usage{
gt_apply_text_decoration(gt_tbl, decoration, locations)
}
\description{
Apply text decoration (bold/italic/underline) to gt locations
}
\keyword{internal}
