% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtx.R
\name{gtx}
\alias{gtx}
\title{GT Extended}
\usage{
gtx(
  .data,
  title = NULL,
  subtitle = NULL,
  source_note = NULL,
  y_group_separator = ">"
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{title}{Table title}

\item{subtitle}{Table subtitle}

\item{source_note}{Table source note}

\item{y_group_separator}{Column separator that defines the table hierarchy.}
}
\value{
GT table
}
\description{
GT Extended
}
