% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_person_record.R
\docType{data}
\name{person_record}
\alias{person_record}
\title{Sample dataset of persons}
\format{
A labelled data frame with 2918 rows and 11 variables:
\describe{
\item{person_id}{Numeric identifier for each person}
\item{sex}{Factor indicating the sex of the person}
\item{age}{Numeric age of the person}
\item{marital_status}{Factor indicating marital status}
\item{employed}{Employment status}
\item{seeing}{Functional difficulty in seeing}
\item{hearing}{Functional difficulty in hearing}
\item{walking}{Functional difficulty in walking}
\item{remembering}{Functional difficulty in remembering}
\item{self_caring}{Functional difficulty in self-caring}
\item{communicating}{Functional difficulty in communicating}
}
}
\usage{
person_record
}
\description{
This is a synthetic dataset containing person information for demonstration purposes.
}
\examples{

person_record
}
\keyword{datasets}
