% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_helpers.R
\name{remove_labels}
\alias{remove_labels}
\title{Remove all labels}
\usage{
remove_labels(data, ...)
}
\arguments{
\item{data}{A data frame or tibble from which to remove all labels.}

\item{...}{A character vector of column names from which to remove labels. If no columns are specified, all labels will be removed.}
}
\value{
A data frame or tibble with all labels removed. If no columns are specified, all labels will be removed.
}
\description{
Remove all labels
}
\examples{
person_record |>
  generate_frequency(
    seeing,
    hearing,
    walking,
    remembering,
    self_caring,
    communicating
  ) |>
  collapse_list() |>
  remove_labels()
}
