% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_helpers.R
\name{rename_label}
\alias{rename_label}
\title{Rename data labels}
\usage{
rename_label(data, ...)
}
\arguments{
\item{...}{A named list of labels to rename. The names should match the column names in the data, and the values should be the new labels.}
}
\description{
Rename data labels
}
