% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg-package.R
\docType{package}
\name{tsg-package}
\alias{tsg}
\alias{tsg-package}
\title{tsg: Generate Publication-Ready Statistical Tables}
\description{
A collection of functions for generating frequency tables and cross-tabulations of categorical variables. The resulting tables can be exported to various formats (Excel, PDF, HTML, etc.) with extensive formatting and layout customization options.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yng-me.github.io/tsg/}
  \item \url{https://github.com/yng-me/tsg}
  \item Report bugs at \url{https://github.com/yng-me/tsg/issues}
}

}
\author{
\strong{Maintainer}: Bhas Abdulsamad \email{aeabdulsamad@gmail.com} (\href{https://orcid.org/0009-0002-5891-8124}{ORCID}) [copyright holder]

}
\keyword{internal}
