% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_crosstab.R
\name{tsg_crosstab}
\alias{tsg_crosstab}
\title{Generate cross-tabulation}
\usage{
tsg_crosstab(
  .data,
  x,
  y,
  x_group = NULL,
  y_group = NULL,
  x_label = tsg_get_config("x_label"),
  y_group_separator = ">",
  x_as_group = FALSE,
  ...
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{x}{\strong{Required}. Row variable to be used as categories.}

\item{y}{\strong{Required}. Column variable.}

\item{x_group}{Row grouping variable/s.}

\item{y_group}{Column grouping variable/s.}

\item{x_label}{Stubhead label (first column).}

\item{y_group_separator}{Column separator that defines the table hierarchy.}

\item{x_as_group}{Use row variable as grouping.}

\item{...}{Accepts valid arguments for \code{tsg_crosstab_inclusion}.}
}
\value{
Returns a cross-table of type \code{tibble}
}
\description{
The function can generate a cross-tabulation that supports more than one (1) variable specified in the column.
}
\examples{

dplyr::starwars |> tsg_crosstab(species, sex)
}
