% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_freq.R
\name{tsg_freq}
\alias{tsg_freq}
\title{tsg_freq}
\usage{
tsg_freq(
  data,
  var_row,
  group_rows = NULL,
  label = NULL,
  code_ref = NULL,
  sort_frequency = F,
  use_var_row_as_group = F,
  ...
)
}
\arguments{
\item{data}{a}

\item{var_row}{b}

\item{group_rows}{c}

\item{label}{d}

\item{code_ref}{e}

\item{sort_frequency}{f}

\item{use_var_row_as_group}{g}

\item{...}{h}
}
\value{
x
}
\description{
tsg_freq
}
