% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_freq_inclusion.R
\name{tsg_freq_inclusion}
\alias{tsg_freq_inclusion}
\title{tsg_freq_inclusion}
\usage{
tsg_freq_inclusion(
  data,
  excluded_cols,
  exclude_total = F,
  exclude_cumulative = F,
  exclude_zero_value = T
)
}
\arguments{
\item{data}{a}

\item{excluded_cols}{b}

\item{exclude_total}{c}

\item{exclude_cumulative}{e}

\item{exclude_zero_value}{d}
}
\description{
tsg_freq_inclusion
}
