% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_frequency.R
\name{tsg_frequency}
\alias{tsg_frequency}
\title{Generate frequency table}
\usage{
tsg_frequency(
  .data,
  x,
  x_group = NULL,
  y_group = NULL,
  x_label = tsg_get_config("x_label"),
  sort_frequency = FALSE,
  use_x_as_group = FALSE,
  ...
)
}
\arguments{
\item{.data}{}

\item{x}{\strong{Required}. Variable to be used as categories.}

\item{x_group}{\strong{Required}. Accepts a vector of string/character as grouping variables.}

\item{y_group}{Column grouping variable/s.}

\item{sort_frequency}{Whether to sort the output. It accepts \code{TRUE} (descending) | \code{FALSE} (ascending). The default is \code{FALSE}}

\item{use_x_as_group}{Use row variable as grouping.}

\item{...}{Accepts valid arguments for \code{tsg_frequency_inclusion}.}

\item{label}{Accepts a string or character value to be used as label for variable or grouping variable.}
}
\value{
Returns a frequency table of type \code{tibble} with optional cumulative total and percent.
}
\description{
Generate a frequency table with optional cumulative total and percent.
}
\examples{
mtcars_cyl_freq <- mtcars |>
  tsg_frequency(cyl)

mtcars_cyl_freq

}
