% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_list.R
\name{tsg_list}
\alias{tsg_list}
\title{tsg_list}
\usage{
tsg_list(
  data,
  list_group,
  var_row,
  ...,
  fn = "tsg_freq",
  list_name_overall = "ALL",
  collapse_overall = F,
  exclude_overall = F
)
}
\arguments{
\item{data}{a}

\item{var_row}{b}

\item{...}{c}

\item{fn}{d}

\item{list_name_overall}{f}

\item{exclude_overall}{g}

\item{list}{e}

\item{label_var_row}{h}

\item{label_list}{i}
}
\description{
tsg_list
}
