% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_multi_response.R
\name{tsg_multi_response}
\alias{tsg_multi_response}
\title{Generate summary table from a multiple response category}
\usage{
tsg_multi_response(
  data,
  var_row,
  var_col,
  group_rows = NULL,
  filter_var = NULL,
  use_var_row_as_group = F
)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{var_row}{\strong{Required}. Row variable to be used as categories.}

\item{var_col}{\strong{Required}. Column variable.}

\item{group_rows}{Column grouping variable/s.}

\item{filter_var}{Use of screening variable.}

\item{use_var_row_as_group}{Use row variable as grouping.}
}
\value{
Returns a cross-table of type \code{tibble}
}
\description{
Generate summary table from a multiple response category
}
