% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_prop.R
\name{tsg_prop}
\alias{tsg_prop}
\title{tsg_prop}
\usage{
tsg_prop(
  data,
  var_row,
  var_col,
  group_rows = NULL,
  group_cols = NULL,
  use_var_row_as_group = F,
  ...
)
}
\arguments{
\item{data}{a}

\item{var_row}{b}

\item{var_col}{c}

\item{group_rows}{d}

\item{group_cols}{e}

\item{use_var_row_as_group}{f}

\item{...}{g}
}
\description{
tsg_prop
}
