% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_prop_inclusion.R
\name{tsg_prop_inclusion}
\alias{tsg_prop_inclusion}
\title{tsg_prop_inclusion}
\usage{
tsg_prop_inclusion(
  data,
  exclude_freq = F,
  exclude_prop = F,
  exclude_total = F,
  prop_label = "Percent"
)
}
\arguments{
\item{data}{a}

\item{exclude_freq}{b}

\item{exclude_prop}{c}

\item{exclude_total}{d}

\item{prop_label}{e}
}
\description{
tsg_prop_inclusion
}
