% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_prop_rename.R
\name{tsg_prop_rename}
\alias{tsg_prop_rename}
\title{tsg_prop_rename}
\usage{
tsg_prop_rename(data, replace_na_with = "Missing")
}
\arguments{
\item{data}{a}

\item{replace_na_with}{b}
}
\description{
tsg_prop_rename
}
