% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_prop_total.R
\name{tsg_prop_total}
\alias{tsg_prop_total}
\title{tsg_prop_total}
\usage{
tsg_prop_total(
  data,
  total_by = "row",
  group_values_by = "statistics",
  format_to_percent = T,
  format_precision = 2,
  total_label = NULL,
  ...
)
}
\arguments{
\item{data}{a}

\item{total_by}{b}

\item{group_values_by}{c}

\item{format_to_percent}{d}

\item{format_precision}{e}

\item{total_label}{f}

\item{...}{g}
}
\description{
tsg_prop_total
}
