% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_prop_total_stack.R
\name{tsg_prop_total_stack}
\alias{tsg_prop_total_stack}
\title{tsg_prop_total_stack}
\usage{
tsg_prop_total_stack(data, ...)
}
\arguments{
\item{data}{a}

\item{...}{b}
}
\description{
tsg_prop_total_stack
}
