% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{tsg_to_gt}
\alias{tsg_to_gt}
\title{Convert a tsg table to a gt object}
\usage{
tsg_to_gt(
  data,
  title = NULL,
  subtitle = NULL,
  source_note = NULL,
  footnotes = NULL,
  names_separator = "__",
  facade = get_tsg_facade(which = "html")
)
}
\arguments{
\item{data}{A \code{tsg} or data frame object.}

\item{title}{Optional title string.}

\item{subtitle}{Optional subtitle string.}

\item{source_note}{Optional source note string.}

\item{footnotes}{Optional character vector of footnotes.}

\item{names_separator}{Column name separator for spanners. Default \code{"__"}.}

\item{facade}{A facade list for styling. Defaults to the global tsg facade.}
}
\value{
A \code{gt_tbl} object.
}
\description{
Convert a tsg table to a gt object
}
\keyword{internal}
