% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_modifiers.R
\name{add_column_total}
\alias{add_column_total}
\title{Add a column total}
\usage{
add_column_total(data, label_total = "Total", ...)
}
\arguments{
\item{data}{A data frame, tibble, or \code{tsg} object to which a column row will be added.}

\item{label_total}{Label for the total column. Default is "Total".}

\item{...}{Additional named arguments to be added as columns alongside the total column.}
}
\value{
The input data frame with an additional column representing the total of each row.
}
\description{
Add a column total
}
\examples{
# Example data frame
df <- data.frame(
 category = c("A", "B", "C"),
 value1 = c(10, 20, 30),
 value2 = c(5, 15, 25)
 )
add_column_total(df)

}
