% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_frequency.R
\name{generate_frequency}
\alias{generate_frequency}
\title{Generate frequency table}
\usage{
generate_frequency(
  .data,
  x,
  x_group = NULL,
  x_label = get_config("x_label"),
  sort_frequency = FALSE,
  x_as_group = FALSE,
  include_total = TRUE,
  include_cumulative = TRUE,
  exclude_zero_value = FALSE
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{x}{\strong{Required}. Variable to be used as categories.}

\item{x_group}{Accepts a vector of string/character as grouping variables.}

\item{x_label}{Stubhead label (first column).}

\item{sort_frequency}{Whether to sort the output. If set to \code{TRUE}, the frequency will be sorted in descending order.}

\item{x_as_group}{Use row variable as grouping.}

\item{include_total}{Whether to include row total.}

\item{include_cumulative}{Whether to cumulative frequencies.}

\item{exclude_zero_value}{Whether to drop categories with zero (0) values}
}
\value{
Returns a frequency table of type \code{tibble}.
}
\description{
Generate a frequency table with optional cumulative total and percent.
}
\examples{
library(palmerpenguins)

# Example 1: Basic usage

generate_frequency(penguins, species)


# Example 2: Add grouping variable and define label for x

penguins |>
  generate_frequency(
    x = sex,
    x_group = 'species',
    x_label = 'Sex'
   )


# Example 3: Add grouping variable, use x as group, and exclude column total

penguins |>
  generate_frequency(
    x = sex,
    x_group = 'species',
    x_as_group = TRUE,
    include_total = FALSE
 )


# Example 4: Exclude cumulative values and sort the output by frequency

penguins |>
  generate_frequency(
   x = species,
   x_label = 'Species',
   sort_frequency = TRUE,
   include_cumulative = FALSE
 )

}
