% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facade.R
\name{get_tsg_facade}
\alias{get_tsg_facade}
\title{Title}
\usage{
get_tsg_facade(facade = "default", which = c("xlsx", "pdf", "html"))
}
\arguments{
\item{facade}{A character string specifying the name of the facade to retrieve. Defaults to "default". The facade is a YAML file that defines the styling and layout of the table}

\item{which}{A character string specifying the format of the facade to retrieve. Options are "xlsx", "pdf", or "html". Defaults to "xlsx".}
}
\description{
Title
}
