% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_helpers.R
\name{rename_label}
\alias{rename_label}
\title{Rename data labels}
\usage{
rename_label(data, ...)
}
\arguments{
\item{data}{A data frame or tibble to rename labels in.}

\item{...}{A named list of labels to rename. The names should match the column names in the data, and the values should be the new labels.}
}
\value{
A data frame or tibble with the specified labels renamed.
}
\description{
Rename data labels
}
\examples{
person_record |>
  generate_frequency(
    seeing,
    hearing,
    walking,
    remembering,
    self_caring,
    communicating
  ) |>
  collapse_list() |>
  rename_label(category = "Functional difficulty")
}
