% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_crosstab.R
\name{tsg_crosstab}
\alias{tsg_crosstab}
\title{Generate cross-tabulation}
\usage{
tsg_crosstab(
  data,
  var_row,
  var_col,
  group_rows = NULL,
  group_cols = NULL,
  use_var_row_as_group = FALSE,
  separator = ">",
  ...
)
}
\arguments{
\item{data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{var_row}{\strong{Required}. Row variable to be used as categories.}

\item{var_col}{\strong{Required}. Column variable.}

\item{group_rows}{Row grouping variable/s.}

\item{group_cols}{Column grouping variable/s.}

\item{use_var_row_as_group}{Use row variable as grouping.}

\item{separator}{Column separator that defines the table hierarchy.}

\item{...}{Accepts valid arguments for \code{tsg_crosstab_inclusion}.}
}
\value{
Returns a cross-table of type \code{tibble}
}
\description{
The function can generate a cross-tabulation that supports more than one (1) variable specified in the column.
}
