% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_crosstab.R
\name{tsg_crosstab}
\alias{tsg_crosstab}
\title{Generate cross-tabulation}
\usage{
tsg_crosstab(
  .data,
  x,
  y = NULL,
  x_group = NULL,
  y_group = NULL,
  x_label = get_config("x_label"),
  y_group_separator = ">",
  x_as_group = FALSE,
  include_frequency = TRUE,
  include_proportion = TRUE,
  include_column_total = TRUE,
  total_by = "row",
  group_values_by = "statistics",
  format_to_percent = TRUE,
  format_precision = 2,
  total_label = NULL,
  ...
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{x}{\strong{Required}. Row variable to be used as categories.}

\item{y}{\strong{Required}. Column variable.}

\item{x_group}{Row grouping variable/s.}

\item{y_group}{Column grouping variable/s.}

\item{x_label}{Stubhead label (first column).}

\item{y_group_separator}{Column separator that defines the table hierarchy.}

\item{x_as_group}{Use row variable as grouping.}

\item{include_frequency}{Whether to include frequency columns.}

\item{include_proportion}{Whether to include proportion/percentage columns.}

\item{include_column_total}{Whether to include column total}

\item{total_by}{Accepts \code{row} | \code{column}. Whether to apply the sum columnwise or rowwise.}

\item{group_values_by}{Accepts \code{statistics} | \code{indicators}.}

\item{format_to_percent}{Whether to format to percent or proportion.}

\item{format_precision}{Specify the precision of rounding the percent or proportion. Default is \code{2}.}

\item{total_label}{Whether to rename the column total.}

\item{...}{Additional arguments}
}
\value{
Returns a cross-table of type \code{tibble}
}
\description{
The function can generate a cross-tabulation that supports more than one (1) variable specified in the column.
}
\examples{

dplyr::starwars |> tsg_crosstab(species, sex)
}
