% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_crosstab_multi_response.R
\name{tsg_crosstab_multi_response}
\alias{tsg_crosstab_multi_response}
\title{Generate summary table from a multiple response category}
\usage{
tsg_crosstab_multi_response(
  .data,
  x,
  ...,
  y = NULL,
  x_group = NULL,
  x_label = tsg_get_config("x_label"),
  use_x_as_group = F
)
}
\arguments{
\item{x}{\strong{Required}. Row variable to be used as categories.}

\item{y}{\strong{Required}. Column variable.}

\item{x_group}{Column grouping variable/s.}

\item{use_x_as_group}{Use row variable as grouping.}

\item{data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{filter_var}{Use of screening variable.}
}
\value{
Returns a cross-table of type \code{tibble}
}
\description{
Generate summary table from a multiple response category
}
