% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_frequency.R
\name{tsg_frequency}
\alias{tsg_frequency}
\title{Generate frequency table}
\usage{
tsg_frequency(
  .data,
  x,
  x_group = NULL,
  x_label = get_config("x_label"),
  sort_frequency = FALSE,
  x_as_group = FALSE,
  include_total = TRUE,
  include_cumulative = TRUE,
  exclude_zero_value = FALSE,
  ...
)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{x}{\strong{Required}. Variable to be used as categories.}

\item{x_group}{Accepts a vector of string/character as grouping variables.}

\item{x_label}{Stubhead label (first column).}

\item{sort_frequency}{Whether to sort the output. If set to \code{TRUE}, the frequency will be sorted in descending order.}

\item{x_as_group}{Use row variable as grouping.}

\item{include_total}{Whether to include row total}

\item{include_cumulative}{Whether to cumulative frequencies}

\item{exclude_zero_value}{Whether to drop categories with zero (0) values}

\item{...}{Additional agruments}
}
\value{
Returns a frequency table of type \code{tibble} with optional cumulative total and percent.
}
\description{
Generate a frequency table with optional cumulative total and percent.
}
\examples{
mtcars_cyl_freq <- mtcars |>
  tsg_frequency(cyl)

mtcars_cyl_freq

}
