% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_list.R
\name{tsg_list}
\alias{tsg_list}
\title{Generate summary tables as a list based on defined grouping or aggregation}
\usage{
tsg_list(
  data,
  list_group,
  var_row,
  ...,
  fn = "tsg_frequency",
  list_name_overall = "ALL",
  exclude_overall = FALSE,
  collapse_overall = FALSE
)
}
\arguments{
\item{data}{\strong{Required}. A data frame, data frame extension (e.g. a tibble), a lazy data frame (e.g. from dbplyr or dtplyr), or Arrow data format.}

\item{list_group}{\strong{Required}. A factor or categorical variable from \code{data} to be used grouping for the list generated.}

\item{var_row}{\strong{Required}. column name of the variable to be used as categories.}

\item{...}{Accepts valid arguments of the selected function in \code{fn}.}

\item{fn}{Accepts \code{tsg_frequency} | \code{tsg_crosstab}. The default value is \code{tsg_frequency}.}

\item{list_name_overall}{Accepts a string that will be used as name/label for the first list. The default value is \code{All}.}

\item{exclude_overall}{Whether to exclude the overall (aggregate) table (first table) in the list.}

\item{collapse_overall}{Whether to conform the structure of the first table with the rest in the list.}
}
\value{
Returns a list of tables aggregated based on values defined in \code{list_group}.
}
\description{
This function allows you to generate summary tables based on a defined grouping or aggregation.
You can optionally return table values in either frequency or proportion/percentage or both.
It uses any \code{tsg_*} valid functions as calculator to generate desired tabulation structure.
}
\examples{
mtcars_by_cyl_freq <- mtcars |>
  tsg_list(list_group = cyl, var_row = am)

mtcars_by_cyl_freq

mtcars_by_cyl_prop <- mtcars |>
  tsg_list(list_group = cyl, var_row = am, gear, fn = 'tsg_crosstab')

mtcars_by_cyl_prop
}
