% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_list.R
\name{tsg_list}
\alias{tsg_list}
\title{tsg_list}
\usage{
tsg_list(
  data,
  var_row,
  ...,
  fn = "tsg_freq",
  list_grouping = NULL,
  list_name_overall = "ALL",
  exclude_overall = F,
  label_var_row = NULL,
  label_list_grouping = NULL
)
}
\arguments{
\item{data}{a}

\item{var_row}{b}

\item{...}{c}

\item{fn}{d}

\item{list_grouping}{e}

\item{list_name_overall}{f}

\item{exclude_overall}{g}

\item{label_var_row}{h}

\item{label_list_grouping}{i}
}
\description{
tsg_list
}
