% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsg_prop.R
\name{tsg_prop}
\alias{tsg_prop}
\title{Title}
\usage{
tsg_prop(
  data,
  var_row,
  var_col,
  group_rows = NULL,
  group_cols = NULL,
  use_var_row_as_group = F,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
